+++ src/main/java/tornadofx/Lib.kt
@@ -12,8 +12,6 @@ import javafx.scene.control.TableView
 import javafx.scene.input.Clipboard
 import javafx.scene.input.ClipboardContent
 import javafx.scene.input.DataFormat
-import tornadofx.Stylesheet.Companion.left
-import tornadofx.Stylesheet.Companion.right
 import java.io.File
 import java.util.function.Predicate
 
@@ -348,13 +346,73 @@ fun Insets.copy(
         vertical?.toDouble() ?: this.bottom,
         horizontal?.toDouble() ?: this.left
 )
-val Insets.horizontal get() = (left+right)/2
-val Insets.vertical get() = (top+bottom)/2
-val Insets.all get() = (left+right+top+bottom)/4
 
+val Insets.horizontal get() = (left + right) / 2
+val Insets.vertical get() = (top + bottom) / 2
+val Insets.all get() = (left + right + top + bottom) / 4
 
 
 fun String.isLong() = toLongOrNull() != null
 fun String.isInt() = toIntOrNull() != null
 fun String.isDouble() = toDoubleOrNull() != null
 fun String.isFloat() = toFloatOrNull() != null
+
+/**
+ * [forEach] with Map.Entree as receiver.
+ */
+inline fun <K, V> Map<K, V>.withEach(action: Map.Entry<K, V>.() -> Unit) = forEach(action)
+
+/**
+ * [forEach] with the element as receiver.
+ */
+inline fun <T> Iterable<T>.withEach(action: T.() -> Unit) = forEach(action)
+
+/**
+ * [forEach] with the element as receiver.
+ */
+inline fun <T> Sequence<T>.withEach(action: T.() -> Unit) = forEach(action)
+
+/**
+ * [forEach] with the element as receiver.
+ */
+inline fun <T> Array<T>.withEach(action: T.() -> Unit) = forEach(action)
+
+/**
+ * [map] with Map.Entree as receiver.
+ */
+inline fun <K, V, R> Map<K, V>.mapEach(action: Map.Entry<K, V>.() -> R) = map(action)
+
+/**
+ * [map] with the element as receiver.
+ */
+inline fun <T, R> Iterable<T>.mapEach(action: T.() -> R) = map(action)
+
+/**
+ * [map] with the element as receiver.
+ */
+fun <T, R> Sequence<T>.mapEach(action: T.() -> R) = map(action)
+
+/**
+ * [map] with the element as receiver.
+ */
+inline fun <T, R> Array<T>.mapEach(action: T.() -> R) = map(action)
+
+/**
+ * [mapTo] with Map.Entree as receiver.
+ */
+inline fun <K, V, R, C : MutableCollection<in R>> Map<K, V>.mapEachTo(destination: C, action: Map.Entry<K, V>.() -> R) = mapTo(destination, action)
+
+/**
+ * [mapTo] with the element as receiver.
+ */
+inline fun <T, R, C : MutableCollection<in R>> Iterable<T>.mapEachTo(destination: C, action: T.() -> R) = mapTo(destination, action)
+
+/**
+ * [mapTo] with the element as receiver.
+ */
+fun <T, R, C : MutableCollection<in R>> Sequence<T>.mapEachTo(destination: C, action: T.() -> R) = mapTo(destination, action)
+
+/**
+ * [mapTo] with the element as receiver.
+ */
+fun <T, R, C : MutableCollection<in R>> Array<T>.mapEachTo(destination: C, action: T.() -> R) = mapTo(destination, action)