+++ src/test/kotlin/tornadofx/testapps/PojoTreeTableColumns.kt
@@ -36,7 +36,7 @@ class PojoTreeTableColumns : View("Pojo Tree Table Columns") {
 
         // Expand the two first levels
         root.isExpanded = true
-        root.children.forEach { it.isExpanded = true }
+        root.children.withEach { isExpanded = true }
         smartResize()
     }
 }